+++ .eslintrc.yml
@@ -1,1 +1,7 @@
+env:
+  mocha: true
+globals:
+  expect: true
 extends: airbnb
+rules:
+  react/jsx-filename-extension: off
