+++ README.md
@@ -2,6 +2,7 @@
 
 [![CircleCI](https://circleci.com/gh/BMBros/TriviaNative/tree/master.svg?style=svg)](https://circleci.com/gh/BMBros/TriviaNative/tree/master)
 
+
 This project was bootstrapped with [Create React Native App](https://github.com/react-community/create-react-native-app).
 
 Below you'll find information about performing common tasks. The most recent version of this guide is available [here](https://github.com/react-community/create-react-native-app/blob/master/react-native-scripts/template/README.md).
