+++ package.json
@@ -8,7 +8,8 @@
     "eject": "react-native-scripts eject",
     "android": "react-native-scripts android",
     "ios": "react-native-scripts ios",
-    "test": "node node_modules/jest/bin/jest.js --watch"
+    "test": "node node_modules/jest/bin/jest.js src",
+    "lint": "eslint src"
   },
   "jest": {
     "preset": "jest-expo"
@@ -16,7 +17,7 @@
   "dependencies": {
     "expo": "^21.0.0",
     "react": "16.0.0-alpha.12",
-    "react-native": "^0.48.4"
+    "react-native": "^0.48.3"
   },
   "devDependencies": {
     "eslint": "^4.9.0",
@@ -25,6 +26,7 @@
     "eslint-plugin-jsx-a11y": "^6.0.2",
     "eslint-plugin-react": "^7.4.0",
     "jest-expo": "^21.0.2",
+    "jest-junit": "^3.1.0",
     "react-native-scripts": "1.5.0",
     "react-test-renderer": "16.0.0-alpha.12"
   }
